docs(examples): 📚 Add link to asciinema preview
Most users expect to be able to click a "play" (▶) icon in order to play the embedded media, however, as this is a thumbnail image the original behavior simply opened the thumbnail, which can be confusing for some. The default behavior after this change will now open the embedded media upon clicking it.